add placeholder help in template editor
